Mosquito Magnet Propane Use
I did a quick repair on my Mosquito Magnet unit. Looked like a loose thermistor issue. I wanted to run some tests to confirm that it is working properly. There are only two tests I can think of: (1) Run in a mosquito-ridden area for a day and verify what I catch, and (2) measure fuel use and confirm it is meeting manufacturer specification.